+++ lib/modules/pool/actions/claim/ClaimModal.tsx
@@ -17,7 +17,7 @@ import { Pool } from '../../usePool'
 import { useClaiming } from './useClaiming'
 import { Address } from 'wagmi'
 import { BalTokenReward } from '@/lib/modules/portfolio/useBalRewards'
-import { ClaimableReward } from '@/lib/modules/portfolio/useClaimableBalances'
+import { ClaimableReward } from '@/lib/modules/portfolio/claim/useClaimableBalances'
 
 type Props = {
   isOpen: boolean
