+++ lib/modules/pool/actions/claim/useClaimStepConfigs.tsx
@@ -2,7 +2,7 @@ import { useHasApprovedRelayer } from '@/lib/modules/relayer/useHasApprovedRelay
 import { useHasMinterApproval } from '@/lib/modules/staking/gauge/useHasMinterApproval'
 import { approveRelayerConfig } from '@/lib/modules/relayer/approveRelayerConfig'
 import { minterApprovalConfig } from '@/lib/modules/staking/gauge/minterApprovalConfig'
-import { ClaimAllRewardsButton } from './ClaimAllRewardsButton'
+import { ClaimAllRewardsButton } from '../../../portfolio/claim/ClaimAllRewardsButton'
 import { Address } from 'viem'
 import { GqlChain, GqlPoolStakingType } from '@/lib/shared/services/api/generated/graphql'
