+++ lib/modules/portfolio/claim/ClaimAllVebalRewardsButton.tsx
@@ -1,4 +1,4 @@
-import { ClaimButton } from './ClaimButton'
+import { ClaimButton } from '../../pool/actions/claim/ClaimButton'
 import { useConstructClaimVeBalRewardsStep } from '@/lib/modules/pool/actions/claim/useConstructClaimVeBalRewardsStep'
 
 export function ClaimAllVebalRewardsButton() {
